Use custom local action for redis container in GHA
In order to make the redis container skippable on Windows (where the
GitHub `services` feature is not available), convert it to a custom
action which launches a redis container.

For now, no health check or ping appears to be necessary, but it could
be added later to the run-redis.sh script if needed.