Add RedisField system for redis-backed classes
`funcx_common.redis` includes the following new helpers for general
use.

Serializers/Deserializers:
- DEFAULT_SERDE (strings)
- INT_SERDE
- JSON_SERDE
- FuncxRedisEnumSerde (needs an enum class as an input to the
  constructor)

Field classes:
- RedisField
- HasRedisFields
- HasRedisFieldsMeta

The RedisField class is used to declare descriptor-style class
attributes which persist to Redis with `hset/hget`, pulling `hname`
and `redis_client` from the owning object. Use `HasRedisFields` class
for inheritance, or directly use `HasRedisFieldsMeta` as a
metaclass (these are equivalent), in order to initialize the
RedisField descriptors on a class.

Additionally, `funcx_common.testing` is a new module to provide
helpers or constants to aid during testing. It is not named `_testing`
or any other private name so that any library using `funcx_common`
could pull out the same values as appropriate.