Handle pubsub get_final_messages without unsub
If the caller does not unsubscribe before calling
`get_final_messages()` on `FuncxRedisPubSub`, the result would be an
infinite generator. Worse, it would ignore the empty queue case and
sit in a tight loop if iterated.

To protect against this possibility, check if there are pending
unsubscriptions in the pubsub before allowing the "final message
consumption". If there are not as many pending_unsubscribe_channels as
there are channels, it is an immediate ValueError.