Add scriv to workflow, including checker workflow
- `scriv create --edit` to add a newsfile
- `scriv collect --version ...` wrapped in `make prepare-release`
- A marker comment in the changelog for insertion of news snippets
- Basic scriv config in `setup.cfg`

The only bit here which gets a little bit gnarly is that we don't have
the package version in a source file, so we can't use a scriv
`literal:` to parse it automatically from there. Maybe the version
should move into there and we should do the read-raw-file tricks in
`setup.py`, leaving `setup.cfg` responsible for everything else.