Set redis heartbeats, associated cleanup
Full changes are described in the changelog fragment. But in summary:
- cleanup the "has a connection" class and rename it
- set `health_check_interval` on the redis.Redis objects
- carve out a path for altering the redis.Redis object if we need it
  in the future

On that last point, there's a `default_redis_connection_factory`
function, and it can be replaced or wrapped as necessary if some
caller *really* needs it.

One of the casualties of the cleanup is the "automatic" error logging,
which would only do anything if you setup a logger on `funcx_common`
anyway. Just allowing errors to propagate and be handled by the
application, with no special logging, is more appropriate to most
contexts and doesn't lose us much of anything.