Fix test setup to allow testing against live S3
1. The S3 bucket name is configurable as a pytest CLI option
2. Several prominent AWS_* env vars are given to tox passenv
3. Tests are setup to skip if boto3 or the test bucket are not
   available, as appropriate

In service of (3), this change also makes it possible to do a "no
deps" install which avoids the non-test extras. For this to work,
`moto` has to move out of the `dev` extra and into a dedicated extra,
as it pulls in boto3 as a dependency.

Some unit tests are rearranged as part of this changeset in order to
more easily share helpers and fixture data.