Fix the live S3 test which disagrees with reality
This is a test which asserts something which *should* be true -- that
trying to read from a nonexistent S3 bucket is an error. However, the
test fails when storing the task, not only when retrieving it. As a
result, the error is not seen -- the task simply looks like it has no
result.