Implement first draft of "messagepack" v0
"messagepack" is the (new) name of the message passing protocol for
use between the forwarder and endpoints. This defines a multiplexer
for various supported protocol versions, called `MessagePacker`, which
can convert message objects to and from bytes.

For now, only protocol version 0 is defined, it is set as the default,
and version detection is a stub (which just returns 0).

Plans for v1, differences from the source `funcx-endpoint` message
definitions, and other details are described in a subpackage-specific
readme file.

So far, only Task and HeartbeatReq have been implemented, with plans
for all of the other necessary message types.