Avoid messagepack messages which can't deserialize
The way the code was structured, deserialization of v0 messages would
apply validation which could be failed by inputs which were accepted
by the constructor.
This reorganizes the exceptions slightly so that validation checks are
done during `__init__`, not only on v0 body loading.

No sophisticated validation framework is added. These are just
validation steps performed which raise `InvalidMessageError` if they
fail.