Fix testsuite for task storage (#9)
* Fix test setup to allow testing against live S3

1. The S3 bucket name is configurable as a pytest CLI option
2. Several prominent AWS_* env vars are given to tox passenv
3. Tests are setup to skip if boto3 or the test bucket are not
   available, as appropriate

In service of (3), this change also makes it possible to do a "no
deps" install which avoids the non-test extras. For this to work,
`moto` has to move out of the `dev` extra and into a dedicated extra,
as it pulls in boto3 as a dependency.

Some unit tests are rearranged as part of this changeset in order to
more easily share helpers and fixture data.

* Fix the live S3 test which disagrees with reality

This is a test which asserts something which *should* be true -- that
trying to read from a nonexistent S3 bucket is an error. However, the
test fails when storing the task, not only when retrieving it. As a
result, the error is not seen -- the task simply looks like it has no
result.

* Test __str__ of StorageException

* Improve tests of invalid S3 storage attributes

Add explicit checks for a number of conditions to the task storage
`_get_from_s3` method, then exercise these checks with a simple test
which ensures that these conditions cause exceptions.

* Test behavior of task storage when ID is unknown

Unknown storage types raise an error. Confirm this.

* Test the S3 storage when data is deleted

Data which is successfully stored but later deleted results in an
error. (As opposed to data which was never successfully stored.)

* Fix github workflows to accord with tox changes

`noredis` is no longer a tox factor, it's now `nodeps`