Enable mypy strict mode (#12)
The primary thing this disallows which is present today is "any
generics", in which a generic type is not given a type parameter and
the type param is therefore implicitly "Any". By way of example,
`Iterator` is not allowed, but `Iterator[str]` or `Iterator[Any]`
are both valid.

In most cases, this is just adding an explicit Any type.