Implement first draft of "messagepack" v0 (#10)
* Implement first draft of "messagepack" v0

"messagepack" is the (new) name of the message passing protocol for
use between the forwarder and endpoints. This defines a multiplexer
for various supported protocol versions, called `MessagePacker`, which
can convert message objects to and from bytes.

For now, only protocol version 0 is defined, it is set as the default,
and version detection is a stub (which just returns 0).

Plans for v1, differences from the source `funcx-endpoint` message
definitions, and other details are described in a subpackage-specific
readme file.

So far, only Task and HeartbeatReq have been implemented, with plans
for all of the other necessary message types.

* Add more messagepack messages and new error type

The new message types are
- EPStatusReport
- Heartbeat
- ManagerStatusReport

These are all fully implemented. Their implementations necessitated
the introduction of a new error type for handling the case of an
invalid message body. Therefore, InvalidMessagePayloadError is added
and is used by all existing message types when the payload is detected
as invalid.

* Implement the ResultsAck message type

This is the last remaining message type for the messagepack protocol.
It is now complete in terms of supported message types.

* Add compatibility tests for messagepack

The compatibility tests only feature one test so far, which ensures
that `Task` messages can be read in both directions by messagepack and
the antecedent implementation in funcx-endpoint

* Avoid messagepack messages which can't deserialize

The way the code was structured, deserialization of v0 messages would
apply validation which could be failed by inputs which were accepted
by the constructor.
This reorganizes the exceptions slightly so that validation checks are
done during `__init__`, not only on v0 body loading.

No sophisticated validation framework is added. These are just
validation steps performed which raise `InvalidMessageError` if they
fail.

* Add remaining messagepack compat tests

These check that messagepack is compatible for encoding and decoding
messages with the funcx-endpoint messages module. The existing test
for Task messages has been expanded into a test per message type.

* Fix duplicate text in messagepack readme