Convert messagepack to use pydantic
All message types are now pydantic models. In order to have a space
for message attributes which are distinct from pydantic fields, an
internal class named `Meta` is used. This is a classic strategy used
in several libraries and frameworks (django, marshmallow), and was
recommended as a workaround in the pydantic issue tracker.
To make defining and using the internal Meta class a little easier, a
couple of helpers are added.

Serialization and deserialization can then be implemented as a model
for the envelope type and using pydantic json dumping and dict
loading. With a dict for the data field in the envelope, we can keep
the same dispatch mechanism as before for choosing a message type and
using it to load the main body of a message.

Update the changelog for the messagepack v1 changes.

Some peculiarities from the transition:

- pydantic.ValidationError can be raised on bad loads, so this
  replaces some of the previously defined excetion classes

- pydantic is lenient about loading UUIDs from strings, bytes, and
  ints, coercing them to UUID where possible -- we could add custom
  validation, but it's simpler to accept this

- pydantic does no special handling of unrecognized fields, and cannot
  be asked to return them to the user. Therefore, in order to log
  warnings on unknown fields, a special pass checking `__fields__` is
  needed

- type annotations must be written such that they can be evaluated at
  runtime by the current python, if they are to be read by pydantic.
  That means using `typing.Dict` rather than `dict` for python<3.9