Remove 'is_system_error' from Result error_details
This field is contentious, and may not be necessary at all since it
should be possible to deduce from the `code` string.

Because the field can always be added later, and adding fields is a
little bit easier than making them optional and phasing them out,
remove it for now.