Make initial changes to message types (#23)
* Make initial changes to message types

Remove task types which are known to be unnecessary in the near
future. Implement the Result message type with some initial values
which can be further refined over time. Add TaskCancel as a stub (it
will likely need fields, but none yet).

* Add TaskCancel.task_id and update tests

Test the TaskCancel message type. Additionally, tweak to one of the
Result message tests to ensure that it behaves correctly on
error_details being present or absent.

* Convert ResultErrorCode to a string, not enum

- Using str will be better for robustness when we have to potentially
  accept new and unknown error codes
- Enums can always be used to provide the values to pass as this
  string later
- This lets us remove all of the code related to de/serializing the
  enum and testing that behavior

* Remove 'is_system_error' from Result error_details

This field is contentious, and may not be necessary at all since it
should be possible to deduce from the `code` string.

Because the field can always be added later, and adding fields is a
little bit easier than making them optional and phasing them out,
remove it for now.