Add default MessagePacker to module constants (#30)
* Add default MessagePacker to module constants

Add a default instance of MessagePacker and expose its pack and unpack
methods as functions of the module. As a result, calling code doesn't
need to manage creation of additional objects. Just import and use the
default. If, in a future version, a non-default MessagePacker is
needed, the design around dispatch and protocol versions is still
accessible.

* Make messagepack test robust to new versions

Rather than checking a literal "1"-byte, check against the
protocol_version used by the test. New versions will therefore be
supported.