Add method for messagepack messages to assert type
This is a simple method for letting messages assert that they are one
of a known set of types and raise a nicely formatted error if the
assertion fails.

Unfortunately, this cannot be expressed in the type system in a clean
way at present. A TypeGuard cannot be defined as a union over a
sequence-type argument to a method.

We don't want to build fragile code which is based on stringized
annotations being ealuated at runtime, so for now this just focuses on
giving us the runtime safety which we'll care about in our
applications. Additional `assert` calls, `cast` calls, or
`type:ignore` comments can still be used to narrow types under type
checking.