Add timestamp list to Result object
Replace the flat `exec_start_ms` and `exec_end_ms` fields on the Result
processer with an optional and extensible list of tuples.  The intent is to
capture the tuple (`state_time` [in ns], `TaskState`, `ActorName`) to reflect
when, what, and who created a status update.  Also add a EXEC_START and
EXEC_END enums to TaskState to reflect the newly captured task transitions.