Add actor name constant.
To track which part of the system is responsible for creating a new status
event we include a new ActorName constant for each entity in the system. This
is used in the status event to record which entity created a status event and
is logged along with the task transition.