Accomodate new black version
Black v24.1 now puts ellipses on the same line as function definitions
in some cases, causing flake8 to emit E704. In formatting issues where
black and flake8 disagree, we side with black.

Black also does this for class definitions. There are two places in the
code base where we use `...` as the body of a class, and both are used
to mean "do nothing," as opposed to "fill this in later (ie, when
overriding)." The `pass` keyword is considered more idiomatic in no-op
situations, so that was chosen here instead of ignoring E701.